[ feat ] service module - 패키지 변경 및 필터링 검색 기능 구현
