--- module-service/src/main/java/module/service/ShowingService.java
@@ -1,9 +1,0 @@
-package module.service;
-
-import java.util.List;
-
-import dto.external.MovieShowingResponse;
-
-public interface ShowingService {
-	List<MovieShowingResponse> getTodayShowing();
-}
