+++ module-service/src/main/java/module/service/showing/ShowingService.java
@@ -1,0 +1,9 @@
+package module.service.showing;
+
+import java.util.List;
+
+import dto.movie.MovieShowingResponse;
+
+public interface ShowingService {
+	List<MovieShowingResponse> getTodayShowing(String title, Long genreId);
+}
